move Player components into own dir
